Add comprehensive usage documentation for StringExtensions in README.md
- Documented all major StringExtensions features, including console output, user input, string cleaning, stop word removal, string utilities, and random string generation.
- Provided clear, categorized code examples for each feature.
- Clarified usage notes and best practices.
- Ensured obsolete or internal methods are not referenced.
- Improved onboarding for new users by showing extension method usage patterns.
